test(doctor): table-drive rules_test.go

+++ internal/doctor/rules_test.go
@@ -214,7 +214,7 @@ func TestEvaluate_OOMKill(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 			signals := &collector.Signals{
-				OOM: &collector.OOMSnapshot{Events: tc.events, Count: uint64(tc.count)},
+				OOM: &collector.OOMSnapshot{Events: tc.events, Count: tc.count},
 			}
 			findings := Evaluate(signals, defaultThresholds())
 			f := findingForSeverity(findings, "oom_kill_occurred", SeverityCritical)
@@ -276,7 +276,7 @@ func TestEvaluate_TCPRetransmitStorm(t *testing.T) {
 				TCP: &collector.TCPSnapshot{
 					RetransmitRate:    tc.retransmitRate,
 					TotalRetransmits:  tc.totalRetransmits,
-					ActiveConnections: uint64(tc.activeConns),
+					ActiveConnections: tc.activeConns,
 					TopRetransmitters: []collector.TCPConnectionEntry{
 						{SrcAddr: "10.0.1.5", SrcPort: 45000, DstAddr: "10.0.1.10", DstPort: 5432, Retransmits: 80},
 					},
